+++ tools/iceworks-cli/command/init/generateNpmName.js
@@ -4,4 +4,4 @@ module.exports = function(name, npmScope) {
   // WebkitTransform -> @ice/webkit-transform
   name = kebabCase(name).replace(/^-/, '');
   return npmScope ? `${npmScope}/${name}` : name;
-}
+};