+++ tools/iceworks-cli/command/start/index.js
@@ -11,6 +11,7 @@ const semver = require('semver');
 const { checkAliInternal } = require('ice-npm-utils');
 const goldlog = require('../../lib/goldlog');
 const checkVersion = require('../../lib/checkVersion');
+const log = require('../../lib/log');
 
 const SERVER_PATH = path.join(userHome, '.iceworks-server');
 
@@ -44,6 +45,7 @@ async function start(options = {}) {
       }
     }
   } catch (error) {
+    log.warn('start warning', error.message);
     downloadAndListen(options);
   }
 }
@@ -51,7 +53,7 @@ async function start(options = {}) {
 function downloadAndListen(options) {
   const child = spawn('node', ['./downloadServer.js'], {
     stdio: ['pipe'],
-    cwd: path.join(__dirname, '../'),
+    cwd: __dirname,
   });
 
   child.stdout.on('data', (data) => {
@@ -117,7 +119,7 @@ async function listen(options) {
 
   child.stderr.on('data', (data) => {
     console.log(data.toString());
-  })
+  });
 
   child.on('error', (error) => {
     failedMsg(error);